% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Part2_RWorldCoding.R
\name{setup.plants}
\alias{setup.plants}
\title{Creates a model of an ecosystem that includes plants, herbivores,
and carnivores on a terrain.}
\usage{
setup.plants(repro, survive, comp.mat, names = NULL)
}
\arguments{
\item{repro}{vector of reproductive probability given to each plant species,
carnivore, or herbivore being modeled.}

\item{survive}{vector of survival probability given to each plant species.}

\item{comp.mat}{matrix of competition probabilities with species'
probabilities as columns and rows such that the matrix matches
the number of species^2.}

\item{names}{vector with names of species used in function.}

\item{setup.plants}{wrapper function that includes all impertinent information
about the plant species being modeled.}

\item{plants}{matrix of ecosystem with species, empty cells, or NAs
as columns and rows.}

\item{timestep}{array of ecosystem timesteps with species, empty cells, or NAs
as columns and rows.}

\item{survive}{function that determines if species survive each timestep.}

\item{reproduce}{function that determines if an indidivual is going to reproduce
and provide young for the next timestep.}

\item{disperse}{function that determines which cell the reproducing individual
will disperse into for the next timestep.}

\item{compete}{function that determines which individual wins a cell when
two individuals attempt to enter into the same cell after reproduction.}

\item{row}{row location of individual in matrix.}

\item{col}{column location of individual in matrix.}

\item{data}{inputs data set being used at the time, such as plants matrix.}

\item{info}{inputs the wrapper information that corresponds to the data called.}

\item{setup.carnivore}{wrapper function that includes all impertinent information
about the carnivore species being modeled.}

\item{eat}{vector sequence of integers that represent timesteps until death since
individual was last sated.}

\item{harvest}{vector of carnivore kill probability for a prey item.}

\item{setup.herbivore}{wrapper function that includes all impertinent information
about the herbivore species being modeled.}

\item{kill}{vector of herbivore kill probability for a plant item.}

\item{snake.matrix}{matrix of ecosystem with numeric inidivuals (snakes),
empty cells, or NAs as columns and rows.}

\item{new.loc.carn}{function that determines the next location that the carnivore
can move to for next timestep.}
}
\value{
list object of the reproductive, survival, and competition probabilities,
and character names given per species.

list object

list object
}
\description{
##Group Work##
Scenario: Snakes, which must thermoregulate across the landscape, eat small herbivores (i.e., rabbits).
The plants that make up the diet of those herbivores must survive and reproduce, as do the herbivores
which act as prey to the surviving and reproducing snakes across a multi-temperature terrain.
TEAM
Hailey Wayment: Herbivores
Megen Kepas: Snakes - temperature regulation
Helen Plylar: Terrain - with a temperature variable
Stephanie Landry: Snakes - as carnivores
}
\details{
To determine if your input data meets requirements of functions,
you can use or manipulate the wrappers \code{setup.plants} and
\code{setup.carnivore} and \code{setup.herbivore}.
}
